Adds Cinder client
This change adds Cinder client in services/volume

* Keeps existing tests for Compute Extensions in tempest/tests/compute
* Copies existing volumes test to tests/volume and refactors them to test the
Cinder API, and adds a BaseVolumesTest class
* Renames the Nova Extensions' volumes_client to volumes_extensions_client
* Adds build_interval and build_timeout parameters in config for Cinder
specific tests
* Renames build_interval and build_timeout environment variables in the
Devstack template file to COMPUTE_BUILD_INTERVAL and COMPUTE_BUILD_TIMEOUT
* Adds volume specific environment variables VOLUME_BUILD_INTERVAL AND
VOLUME_BUILD_TIMEOUT to the Devstack template file.

Fixes LP Bug #1026190

+++ tempest/tests/volume/test_volumes_get.py
@@ -18,41 +18,40 @@
 from nose.plugins.attrib import attr
 
 from tempest.common.utils.data_utils import rand_name
-from tempest.tests.compute.base import BaseComputeTest
+from tempest.tests.volume.base import BaseVolumeTest
 
 
-class VolumesGetTest(BaseComputeTest):
+class VolumesGetTest(BaseVolumeTest):
 
     @classmethod
     def setUpClass(cls):
         super(VolumesGetTest, cls).setUpClass()
-        cls.client = cls.volumes_extensions_client
+        cls.client = cls.volumes_client
 
     @attr(type='smoke')
     def test_volume_create_get_delete(self):
-        """CREATE, GET, DELETE Volume"""
+        """Create a volume, Get it's details and Delete the volume"""
         try:
+            volume = {}
             v_name = rand_name('Volume-')
             metadata = {'Type': 'work'}
-            #Create volume
+            #Create a volume
             resp, volume = self.client.create_volume(size=1,
                                                      display_name=v_name,
                                                      metadata=metadata)
             self.assertEqual(200, resp.status)
             self.assertTrue('id' in volume)
-            self.assertTrue('displayName' in volume)
-            self.assertEqual(volume['displayName'], v_name,
+            self.assertTrue('display_name' in volume)
+            self.assertEqual(volume['display_name'], v_name,
             "The created volume name is not equal to the requested name")
             self.assertTrue(volume['id'] is not None,
             "Field volume id is empty or not found.")
-            #Wait for Volume status to become ACTIVE
             self.client.wait_for_volume_status(volume['id'], 'available')
-            #GET Volume
+            # Get Volume information
             resp, fetched_volume = self.client.get_volume(volume['id'])
             self.assertEqual(200, resp.status)
-            #Verfication of details of fetched Volume
             self.assertEqual(v_name,
-                             fetched_volume['displayName'],
+                             fetched_volume['display_name'],
                              'The fetched Volume is different '
                              'from the created Volume')
             self.assertEqual(volume['id'],
@@ -63,33 +62,38 @@ class VolumesGetTest(BaseComputeTest):
                              fetched_volume['metadata'],
                              'The fetched Volume is different '
                              'from the created Volume')
-
+        except:
+            self.fail("Could not create a volume")
         finally:
-            #Delete the Volume created in this method
-            resp, _ = self.client.delete_volume(volume['id'])
-            self.assertEqual(202, resp.status)
-            #Checking if the deleted Volume still exists
-            self.client.wait_for_resource_deletion(volume['id'])
+            if volume:
+                # Delete the Volume if it was created
+                resp, _ = self.client.delete_volume(volume['id'])
+                self.assertEqual(202, resp.status)
+                self.client.wait_for_resource_deletion(volume['id'])
 
     @attr(type='positive')
     def test_volume_get_metadata_none(self):
-        """CREATE, GET empty metadata dict"""
+        """Create a volume without passing metadata, get details, and delete"""
         try:
+            volume = {}
             v_name = rand_name('Volume-')
-            #Create volume
+            # Create a volume without metadata
             resp, volume = self.client.create_volume(size=1,
                                                      display_name=v_name,
                                                      metadata={})
             self.assertEqual(200, resp.status)
             self.assertTrue('id' in volume)
-            self.assertTrue('displayName' in volume)
-            #Wait for Volume status to become ACTIVE
+            self.assertTrue('display_name' in volume)
             self.client.wait_for_volume_status(volume['id'], 'available')
             #GET Volume
             resp, fetched_volume = self.client.get_volume(volume['id'])
             self.assertEqual(200, resp.status)
             self.assertEqual(fetched_volume['metadata'], {})
+        except:
+            self.fail("Could not get volume metadata")
         finally:
-            #Delete the Volume created in this method
-            resp, _ = self.client.delete_volume(volume['id'])
-            self.assertEqual(202, resp.status)
+            if volume:
+                # Delete the Volume if it was created
+                resp, _ = self.client.delete_volume(volume['id'])
+                self.assertEqual(202, resp.status)
+                self.client.wait_for_resource_deletion(volume['id'])
